pyproject.toml
[tool.poetry]
name = "arrowmancer-gym"
version = "0.1.0"
description = "RL environment for Spellbrush's Arrowmancer"
authors = ["Cooper Miller <kcoopermiller9@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gymnasium = "^0.29.1"
torchrl = "^0.4.0"
tqdm = "^4.66.4"
pygame = "^2.5.2"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
transformers = "^4.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
